ref(uptime): Reuse UptimeChecksGrid between issues / details
